pyproject.toml
[project]
name = "telegram-reader-mcp"
version = "0.3.1"
description = "Telegram MCP Server - 让 Claude 读取 Telegram 消息"
requires-python = ">=3.11"
dependencies = [
    "telethon>=1.34.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.0.0",
    "aiosqlite>=0.19.0",
]

[project.scripts]
telegram-mcp = "telegram_mcp.server:main"
telegram-mcp-login = "telegram_mcp.cli:login"
telegram-mcp-status = "telegram_mcp.cli:status"
telegram-mcp-cache-stats = "telegram_mcp.cli:cache_stats"
telegram-mcp-cache-clear = "telegram_mcp.cli:cache_clear"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/telegram_mcp"]

[tool.uv]
dev-dependencies = []
